Fully working paste
